replace version manager v1 with v2
