+++ storage/src/lsm_tree/manifest/mod.rs
@@ -1,6 +1,5 @@
 mod error;
 mod version;
-pub mod version_v2;
 
 pub use error::*;
 pub use version::*;
